qman: Add enqueue api variant
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

Add a QMan enqueue api variant which invokes a user supplied callback function
just before writing the commit verb. The callback function is invoked
when the request is guaranteed to be enqueueud. The standard enqueue request
can return an error. Useful when the caller is saving state information
and isn't able to roll back the transaction.

Signed-off-by: Jeffrey Ladouceur <Jeffrey.Ladouceur@freescale.com>
Change-Id: I832dbb581758d7cae63e2d10c88b2e9f5c0bb465
Reviewed-on: http://git.am.freescale.net:8181/4291
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Thorpe Geoff-R01361 <Geoff.Thorpe@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>