dpaa_eth: Add FQ taildrop threshold for mac-less devices
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

FQ taildrop threshold is set by each partition on its RX frame queues
which are the TX queues of the other partition.
It is safe to rely on one partition to set the FQ taildrop threshold
on behalf of the other partition because the ERN notifications will
be delivered to the portal enqueueing the frames.

Change-Id: I6a3ab6511a7e2bbe293cc68ddae4d17e075cf4fa
Signed-off-by: Cristian Sovaiala <cristian.sovaiala@freescale.com>
(cherry picked from commit 1c20f0e8cb943f31eb2ad8c626a4c9fd9c8421b6)
Reviewed-on: http://git.am.freescale.net:8181/1055
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>