powerpc/fsl_book3e: set the correct physical id for the secondary thread in cpu hotplug
For the first booting, we would tweak the PIR register to get a
sequential physical id for all the threads. But in cpu hotplug,
the core will be reset when all the threads in a core are disabled.
That means the PIR register may contains the correct cpu physical id
or the hardware initialized value. And we don't have a way to
distinguish these two cases. We also can't set the cpu physical id
to a general register as what the kexec/kdump do since this register
will also be initialized after the core reset.

So we choose to get the cpu physical id from the PACA struct. Before
that we need to know the cpu logical id first. Fortunately we can
get this based on the current thread cpu logical id and TIR register.
This method also works for the booting.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>