dpa_offload: Automatically enable CRC8 when DPA offloading drivers are enabled
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

DPA offloading drivers are dependent on the CRC8 functions.
If one enables the DPA offloading drivers for build, but forgets
to enable CRC8, the Linux kernel build will fail in the link
stage.

This change automatically selects CRC8 for build when the user
wants DPA offloading driver in the kernel.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Ife1bb645056169a0337f9d1f6318a39862ec3ee6
Reviewed-on: http://git.am.freescale.net:8181/2482
Reviewed-by: Zanoschi Aurelian-B43522 <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>