dpa_ipsec: Fix support for Policer engine
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

The support for Policer engine was either lacking
or had multiple bugs. A Policer engine can be
configured in multiple cases: policy miss action,
post ipsec action and/or policy action.  A clean
fix-up could only be performed if the code of
non-compat mode is completely separated from the
code of compat mode.

This patch is related to CR ENGR00260484.

Signed-off-by: Anca Jeanina FLOAREA <anca.floarea@freescale.com>
Change-Id: Iecb43d7efddad550ee25a982150e656ca5216c2d
Reviewed-on: http://git.am.freescale.net:8181/2557
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>