dpaa_eth: Increase maximum allowed size for buffer recycle
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

There is an upper limit for the size of a recycleable skb, in order
to prevent unbounded growth of the recycled buffers. This limit
was however too strict and could prevent recycling for some of the
skbs found on the ipforward path.

Relax the recycling condition, such as to also account for internal
buffer padding, skb alignment padding and sh_info size.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I0f0b8a48a564657dd107661d6bb5ad474ce8a7a8
Reviewed-on: http://git.am.freescale.net:8181/2807
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Hamciuc Bogdan-BHAMCIU1 <bogdan.hamciuc@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>