fmd: Update range for FSL_FM_RX_EXTRA_HEADROOM
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

We need to reserve some space in the Rx skb buffer for NET_IP_ALIGN,
otherwise on forwarding the stack may reallocate the skb header.
The most easily available place to do this is the Rx extra headroom,
so enforce the minimum non-zero possible value as lower end of the
configuration range.

While here, enforce (and comment on) the restriction to only set values
which are a multiple of 16 bytes, which was missed in the previous
version of the help text.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I9403a36f11cc1c7b9460126fb5ab97237f9bd105
Reviewed-on: http://git.am.freescale.net:8181/3482
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Hamciuc Bogdan-BHAMCIU1 <bogdan.hamciuc@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>