dpaa_eth: Fix computation of FD headroom
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

The Rx/Tx headrooms are aligned to a configurable value; if that value is
zero however, FMD will interpret it as a "don't care" and use its own
internal default, which is currently 64.
If the user configures a zero headroom alignment, use our own default,
otherwise we risk getting out of sync with the actual hardware configuration.

Also, don't assume the Rx headroom is always equal to the Tx headroom plus
the Kconfigurable Rx extra headroom, since alignment may change this.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: Iaf12deb55f4330edb46c6d358114f8f757b64f08
Reviewed-on: http://git.am.freescale.net:8181/3479
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Hamciuc Bogdan-BHAMCIU1 <bogdan.hamciuc@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Sovaiala Cristian-Constantin-B39531 <Cristian.Sovaiala@freescale.com>
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>