dpaa_eth: Fix freeing of frame in error cases
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

In case of a Tx error at FMan level or an ERN, if the frame was
marked as recycleable we need to put the buffers in the target pools,
otherwise we mess up the pool counters. This was done correctly for
the ERN callback, now fixing the Tx error one as well.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Change-Id: I0faa77a31af9086806713e593da72b03c2bd3223
Reviewed-on: http://git.am.freescale.net:8181/4192
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bucur Madalin-Cristian-B32716 <madalin.bucur@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>