Qman/kexec: use qm_drain_dqrr
When we rebase Qman driver on SDK v1.4, one previous commit,

"qman: drain the dqrr entries in kexec boot"

is changed, and qm_drain_dqrr() is replaced with current
qm_dqrr_cdc_consume_n().

But it seems that qm_dqrr_cdc_consume_n() only mark all
the DQRR entries as being consumed in the register. But it
doesn't update the corresponding ci, cursor, fill and vbit
in struct qm_dqrr, this will cause inconsistency between
software and hardware. When we use these values in ci and
vbit to get the current FD from DQRR entries, the content of
it is definitely stale. So we should use the qm_drain_dqrr
to drain the DQRR entries here.

Signed-off-by: Tiejun Chen <tiejun.chen@windriver.com>
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>