dpa_offload: Leave stats counters Ids memory allocated for asynchronous statistics requests
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

When the DPA Stats is processing asynchronous requests, the memory allocated
for the counter Ids and fed to the processing tasklet must NOT be released
in the get_counters call, because the tasklet would work on invalid memory
space later on. This memory must be released only when the event which
signals the completion of the async request is consumed by the user space.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I49481a244305fb468781d3f3b1305698b554055f
Reviewed-on: http://git.am.freescale.net:8181/4562
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zanoschi Aurelian-B43522 <Aurelian.Zanoschi@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>