powerpc64: calculate the count of rela entries from RELASZ and RELAENT
In the current kernel, we get the rela entries count from the
RELACOUNT entry in the dynamic table. But this entry only count for
the R_PPC64_RELATIVE relocation type. So it will not work if there are
other relocation types in the .rela.dyn section. We choose to calculate
the real count of rela entries by using the RELASZ and RELAENT entries
in the dynamic table.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>