fsl_qbman: Make local symbols 'static'
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

This fixes the following sparse warning(s)

warning: symbol 'bman_portal_max' was not declared. Should it be static?
warning: symbol 'params_nocb' was not declared. Should it be static?
warning: symbol 'params_cb' was not declared. Should it be static?
warning: symbol 'qman_ceetm_configure_lfqmt' was not declared. Should it be static?
warning: symbol 'qman_ceetm_query_lfqmt' was not declared. Should it be static?
warning: symbol 'qman_ceetm_configure_cq' was not declared. Should it be static?
warning: symbol 'qman_ceetm_configure_dct' was not declared. Should it be static?
warning: symbol 'qman_ceetm_query_dct' was not declared. Should it be static?
warning: symbol 'qman_ceetm_configure_class_scheduler' was not declared.  Should it be static?
warning: symbol 'qman_ceetm_query_class_scheduler' was not declared.  Should it be static?
warning: symbol 'qman_ceetm_configure_mapping_shaper_tcfc' was not declared. Should it be static?
warning: symbol 'qman_ceetm_query_mapping_shaper_tcfc' was not declared.  Should it be static?
warning: symbol 'qman_ceetm_configure_ccgr' was not declared. Should it be static?
warning: symbol 'qman_ceetm_cq_peek_pop_xsfdrread' was not declared.  Should it be static?
warning: symbol 'qman_ceetm_query_statistics' was not declared. Should it be static?
warning: symbol 'qman_ceetm_write_statistics' was not declared. Should it be static?
warning: symbol 'qman_get_ip_revision' was not declared. Should it be static?

Signed-off-by: Emil Medve <Emilian.Medve@Freescale.com>
Change-Id: I167a9f0d7851a05150be000a927daaac4ca03482
Reviewed-on: http://git.am.freescale.net:8181/8100
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Haiying Wang <Haiying.Wang@freescale.com>
Reviewed-by: Emilian Medve <Emilian.Medve@freescale.com>