powerpc/kgdb: Fix gdb fails to get PC register in e500mc
gdb can't single-step since it fails to get PC.

kgdb send message including PC to gdb, but the address to store PC is
wrong. So gdb fails to get it.

In e500mc, each FP register is 64-bit nor 32-bit which is only used for
e500 SPE.
In message sent to gdb, the PC is stored behind FP. So the wrong size of
FP makes PC stored in wrong address.

Signed-off-by: Bin Jiang <bin.jiang@windriver.com>