dpa_offload: Add SA per DSCP support in DPA IPSec
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

If different classes of traffic defined by DSCP are sent on the same SA
and if the receiver is using an optional anti-replay feature, this could
result in discarding lower priority packets. To avoid this, the sender
should send traffic with different classes, but the same selector values
on different SA's in order to support QoS appropriately. The IPSec
implementation must permit establishment and maitenance of multuple SA's
between sender and receiver, with the same selectors. The receiver must
process the packets from the different SAs without prejudice. These
requirements apply to both transport and tunnel mode SA's. (RFC-4301)

On IPSec inbound path there are no needed modifcations in order to
support this feature. The idea behind this feature is to direct outbount
traffic with different DSCP to different outbound SA's, even if it has
the same SA selector with the rest of the traffic.

On the outbound direction the selectors for the policies are partially
custom. Now the user can select fields from the following group of
supported fields:

	* IPSA (supports mask)
	* IPDA (supports mask)
	* IPPROTO
	* DSCP
	* SPORT (for TCP/UDP/SCTP)
	* DPORT (for TCP/UDP/SCTP)
	* ICMP_TYPE
	* ICMP_CODE

The format of the key is fixed and cannot be modified, but the user has
the possibility to mask the fields wants to use in the policy selector.
For the DSCP field, the user can set a range of values between
dscp_start and dscp_end. The DPA IPSec will create a policy selector for
every DSCP value entered by the user. In case dscp_start equals
dscp_end, only one policy will be added.
This requires the user to properly configure the outbound pre-SEC CC
nodes to generate the key using also the DSCP field.
The statistics returned for the SA with a range of DSCP values defined,
will totalize the statistics for every policy selector.

Change-Id: I96c1629ca2d52bc6023af494929e27865348230a
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/9874
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Andrei Varvara <andrei.varvara@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>