spi: zynq-qspi: Check for busy in setup and call setup_transfer
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

In setup, call setup transfer to ensure clock configuration is done
before spi is enabled.
Check for busy so as to not interrupt existing transfer with this config.
This is necessary as clock config needs to happen before when disabled.
The current transfer_one hook used will not allow the earlier flow of
disabling and enabling.

Signed-off-by: Harini Katakam <harinik@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 2103345769e5cecdbb4000be29afd79b6c13cfcc)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>