gianfar: Add BD rings allocation to L2 SRAM
When this option is enabled the driver will try to allocate
the buffer descriptors (BDs) to the L2 SRAM memory, using
the powerpc/85xx cache-sram support API for the P1/P2 QorIQ
platforms.  For this to work, the size and address of the
SRAM region must be provided as kernel boot params.
Should the attempt to allocate in L2 SRAM fail, the driver
falls back to normal allocation of BDs as before (to DDR).
The gfar_l2sram_en driver module param is provided to allow
to disable the allocation to L2 SRAM at driver level
(enabled by default).

Change-Id: I087e97baf289111a0841fdf69592c56d3a9a0310
Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11624
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jianhua Xie <jianhua.xie@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
[Xulin: Original patch taken from
QorIQ-SDK-V1.7-SOURCE-20141218-yocto_RDS_20150206.iso]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>