powerpc/booke64: make sure the secondary thread get the correct physical id
The commit ("powerpc/fsl_book3e: set the correct physical id for the
secondary thread in cpu hotplug") change the method which we use to set
the physical id for the secondary thread. First we will store the
physical id of the first thread to a global variable __thread0_cpu_id,
and then release the secondary thread. The secondary thread will use
this value to calculate its physical id. So we must make sure that
the store to __thread0_cpu_id is performed to the secondary thread
before the instruction of the releasing of secondary thread. Add
a 'sync' for this purpose.

Reported-by: Yong Zhang <yong.zhang@windriver.com>
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>