dpa_offload: Fix SA per DSCP IPv6 support
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

In order to extract the Traffic Class in case of IPv6 flows, the keygen
will add two bytes to the key which hold: IPv6 version (4 bits), TC (8
bits) and 4 bits zero from Flow Label. At the time the feature was
developed this information wasn't available and the assumption that it
works as IPv4 case was made.

Change-Id: Iedfab5635ba8a8c1a15404dc830edfe60844ada6
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11712
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>