gianfar: Use netif_device_{detach,attach}() around reset
We need to keep the TX queues stopped throughout a reset,
without triggering the TX watchdog and regardless of the
link state.  Given the way the watchdog works, the proper
way to do this is to use netif_device_{detach,attach}()
just as we do around suspend/resume.
netif_device_detach() can be called to stop TX queues in a
control operation other than ndo_stop, and in our case
this is insured by testing the driver's RESETTING state
flag.
napi_disable() is called before the stopping the TX queues
to make sure that the poll function is not running,
and therefore there can be no netif_wake_queue() calls.
(Inspired by sfc.)

Change-Id: Ie00ddaf39e454938b4ab5418d038de5d1a277b03
Signed-off-by: Claudiu Manoil <claudiu.manoil@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/11994
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Rajan Gupta <rajan.gupta@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
[Xulin: Original patch taken from
QorIQ-SDK-V1.7-SOURCE-20141218-yocto_RDS_20150206.iso]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>