asf_gfar: Patch to fix the crash seen in ifup for bsc9132.
ASF uses two TX queues to avoid contention. BSC9132 even though having
controller which can support two groups was having the DTS file written
for one group. To fix this we allow ASF to override the default tx queue
mapping with the "fsl,tx-bit-map" value given in the DT.

This patch resolves the crash seen in case of bsc9132 board
when the network interface is up and trying to xmit the packet.

Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>
[Xulin: Original patch taken from
QorIQ-SDK-V1.7-SOURCE-20141218-yocto_RDS_20150206.iso]
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>