crypto: caam/qi - fix handling of create_caam_req_fq(...) error
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

When create_caam_req_fq(...) returns an error, it does so by
using the ERR_PTR macro, due to its return type.
Thus the return of this function needs to be checked by using
the corresponding IS_ERR_* class of macros, and not by
a "if(!ret)" construct.
This patch adds the necessary macro calls in all the locations
where create_caam_req_fq(...) is called.

Change-Id: I41283fc1f728d959099351cc36ba07ba377a1848
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/12066
Reviewed-by: Mircea Pop <mircea.pop@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Horia Ioan Geanta Neag <horia.geanta@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>