fmd: enable addition of CCtable entry not linked to a keygen
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

This fixes the problem that if at a time, a path between a keygen and
a CCtable is broken (e.g. by deleting an entry in a previous table)
then it is impossible to add any entry in that table or in any tables
that are "behind" it until a new path would be created

Also fixing the case w.r.t. to share CC nodes when
deleting entry in a CC node attached to first port will corrupt
context of ccnode shared with other ports.

Signed-off-by: Mandy Lavi <mandy.lavi@freescale.com>
Change-Id: I089982acee71c489b6ed9275b641afa74315c18f
Reviewed-on: http://git.am.freescale.net:8181/12820
Reviewed-by: Nir Erez <nir.erez@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Tested-by: Jose Rivera <German.Rivera@freescale.com>
(cherry picked from commit e25a50bbcd56f0c1f4a725b70eab5f75604d14a3)
Reviewed-on: http://git.am.freescale.net:8181/13059