ARM: zynq: Implement better check for DDR idle
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

In suspend DRAM is put in self-refresh and clocks are stopped. Though,
before the clocks can be stopped the DDRC must be idle. Zynq does not
provide a reliable way to query the activity state of the DRAM. The
recommened way is to poll the CAM queues and wait for ~40 DDR cycles
once they are idle to allow the queues and pipelines to be drained.

If something causes new DDR transactions in the gap between finish
polling and turning DDR off, things will go sideways...

Tested-by: Chandramohan Pujari <cpujari@xilinx.com>
Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit a5f3bac3c39aa0a218d0ba59947b04b7ec6da624)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>