onic: increase frequency of draining
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

The draining function will be executed when each CPU has added more than 8
buffers in the draining pool. In some conditions, the sock buffers waits for
the Ethernet driver to release the buffer in order to continue its
transmission.

To avoid memory depletion in the sock buffers, the Ethernet driver will release
faster the received skb from the stack.

This patch makes onic to call the drain procedure on RX path also even if there
is only one buffer in the bpool.

Change-Id: I19f800bbd17c23fda5739b7613fab3da355c6fbc
Signed-off-by: Marian Rotariu <marian.rotariu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/13095
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>