net: xilinx_emacps: Make MTU of 1500 possible
The current driver limits the MTU to  (1500 - hard_header_len). If the kernel
boots the MTU is set to 1500 but if once changed to < 1500 it is not possible
to set it back to 1500 again.

After this fix it is possible to set the MTU back to 1500.

The specific change MTU function is replaced by the standard eth_change_mtu
one.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@netmodule.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit dc0c3868b50df4caaeedf85b29359180d7a1197d)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>