Add support for hardware threads on e6500.
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

The general idea is that each core will release all of its
threads into the secondary thread startup code, which will
eventually wait in the secondary core holding area, for the
appropriate bit in the PACA to be set. The kick_cpu function
pointer will set that bit in the PACA, and thus "release"
the core/thread to boot. We also need to do a few things that
U-Boot normally does for CPUs (like enable branch prediction).

Signed-off-by: Andy Fleming <afleming@freescale.com>

Signed-off-by: Zhao Chenhui <chenhui.zhao@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>
[some bug fixes for hotplug support]

Signed-off-by: Scott Wood <scottwood@freescale.com>
[Fixed feature section placement]
Signed-off-by: Lu.Jiang <lu.jiang@windriver.com>