fsl/powerpc: fix kernel cannot boot under topaz
[Original patch taken from QorIQ-SDK-V1.6-SOURCE-20140619-yocto.iso]

when we have embedded hypervisor functionality, the MMUCFG[LPIDSIZE]
is not zero, this is issue will occur.

asm functions (setup_altivec_idle + setup_pw20_idle) that overwrite
register r4. Register r4 holds the second parameter of the
setup___setup_cpu_e* functions and is a pointer to the cpu spec features.
Under baremetal the param is not used so the problem doesn't show up,
but under hypervisor it is used.

Signed-off-by: Wang Dongsheng <dongsheng.wang@freescale.com>
Tested-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>
Change-Id: I0fe5f7814ff0a77988f3b8540933a522f3217679
Reviewed-on: http://git.am.freescale.net:8181/4789
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Tudor Laurentiu-B10716 <Laurentiu.Tudor@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>
Signed-off-by: Lu.Jiang <lu.jiang@windriver.com>