offline_port: fix dt parser for tx frame queues
With the current implementation the user has the possibility to initialize
some queues that exit the offline port. The format of the device tree format
is:

 fsl,qman-frame-queues-egress = <base_id1 count1 ... base_idn countn>;
 fsl,qman-channel-ids-egress = <channel_id1 ... channel_idn>;

Intuitively, the base_id1, count1 frame queues batch are placed in the
channel_id1 QMan channel. Because the list_add() adds an entry at the beginning
of the list, the map between the frame queues and their channel ID is reversed.

This patch adds a particular entry at the end of the fq list using
list_add_tail() function.

Signed-off-by: Marian Cristian Rotariu <marian.rotariu@freescale.com>
Change-Id: I8d3a2fd3c36f63d15837ffe90bf6e4e8b041a3e5
Reviewed-on: http://git.am.freescale.net:8181/20154
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Tested-by: Marian Cristian Rotariu <marian.rotariu@freescale.com>
[Lu:Original patch taken from
QorIQ-SDK-V1.7-SOURCE-20141218-yocto.iso]
Signed-off-by: Jiang Lu <lu.jiang@windriver.com>