dma: make fsldma depend on PPC
The fsldma driver is only buildable on ppc platform. Previously it has
a dependency on FSL_SOC. This was fine since FSL_SOC was only enabled
on some ppc boards. But with the introduction of fsl ls boards, the
FSL_SOC may be enabled on a arm platform and would cause build error
if this driver is enabled. So make it also depend on PPC.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>