net: fsl: make ucc_geth depend on PPC
The ucc_geth driver is only buildable on ppc platform. Previously it has
a dependency on QUICC_ENGINE. This was fine since QUICC_ENGINE was only
enabled on some ppc boards. But with the introduction of fsl ls boards,
the QUICC_ENGINE may be enabled on a arm platform and would cause build
error if this driver is enabled. So make it also depend on PPC.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>