ls1021a/pci: add LS1021A PCIe Power Management support
This patch only for LS1021A PCIe Power Managment, because there has
a PCIe hardware issue, we need a workaround to support it.

LS1021A Workaround for internal TKT228622 to fix the INTx hang issue.

If this hardware be fixed, we not need to do anything for PCIe PM.

Signed-off-by: Wang Dongsheng <dongsheng.wang@freescale.com>
Change-Id: I7b689e4780322f77bebd8671da71d82f6fdb18a3
Reviewed-on: http://git.am.freescale.net:8181/22585
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Minghuan Lian <Minghuan.Lian@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>
[Lu:Original patch taken from
QorIQ-SDK-V1.7-SOURCE-20141218-yocto.iso]
Signed-off-by: Jiang Lu <lu.jiang@windriver.com>